Safari-611.1.10/LayoutTests/http/tests/media/video-auth-with-allowCrossOriginSubresourcesToAskForCredentials-expected.txt

http://127.0.0.1:8000/media/resources/video-auth.php?name=test.mp4&type=video/mp4 - didReceiveAuthenticationChallenge - Responding with username:password
http://localhost:8000/media/resources/video-auth.php?name=test.mp4&type=video/mp4 - didReceiveAuthenticationChallenge - Responding with username:password

Tests that the media player sends authorization credentials when requesting a media file.
Testing same domain (127.0.0.1)
EVENT(canplay)
Testing cross domain (localhost)
EVENT(canplay)
END OF TEST
